drf_social/requirements.txt
django-allauth
django-rest-auth